+++ Web/Pages/Transactions/Enum.cs
@@ -1,0 +1,12 @@
+﻿using System.ComponentModel;
+
+namespace Web.Pages.Transactions
+{
+    public enum TransactionTypeGroup
+    {
+        [Description("Debet")]
+        Debit,
+        [Description("Credit")]
+        Credit
+    }
+}
